production_backend/requirements.txt
# Production requirements for the building footprint backend server

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
httpx==0.25.2

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.7
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7

# AWS Integration
boto3==1.34.0
botocore==1.34.0
watchtower==3.0.1

# Task Queue
celery[redis]==5.3.4
redis>=4.5.2,<5.0.0

# Data Processing
numpy==1.24.3
pandas==2.0.3
shapely==2.0.2
geopandas==0.14.1
rasterio==1.3.9

# ML Dependencies (optional - can be installed separately)
torch==2.1.1
torchvision==0.16.1
opencv-python-headless==4.8.1.78

# Configuration
pydantic==2.5.0
pydantic-settings==2.0.3
python-dotenv==1.0.0

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0
prometheus-client==0.19.0

# Development Tools
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Utilities
click==8.1.7
rich==13.7.0
typer==0.9.0
python-dateutil==2.8.2
requests==2.31.0
aiofiles==23.2.1